Describe an approach to use integration tests instead of unit tests (#671)
In some situations integration tests can be used instead of unit tests to provide full test coverage. Usually integration tests are used to verify that modules of a software system work together as designed and unit tests cover the functionality of isolated units of the system. In the approach described in this PR the test coverage of integration tests is extended to verify the correct behaviour of the system as well.

Resolves #669

Add a paragraph about an approach for using integration tests instead of unit tests into “Unit testing” document in the playbook.
Changes were made in "Unit testing" document: docs/automated-testing/unit-testing/README.md